🐛 Bug(email): fix partner email when deleting an account that don't have any name (#886)

+++ dossierfacile-api-tenant/src/main/java/fr/dossierfacile/api/front/service/UserServiceImpl.java
@@ -122,7 +122,9 @@ public class UserServiceImpl implements UserService {
 
         TransactionalUtil.afterCommit(() -> {
             mailService.sendEmailAccountDeleted(tenantToDeleteDto);
-            keycloakService.deleteKeycloakUserById(keycloakId);
+            if (StringUtils.isNotBlank(keycloakId)) {
+                keycloakService.deleteKeycloakUserById(keycloakId);
+            }
         });
     }
